Added a pause menu
Pressing tab (because escape is already taken) brings up a pause menu. You can resume or quit the game. It will freeze the enemies AI, cache their velocity and reset it to 0. Then on resume, it will set their velocities back to what they were before pausing. There are a few issues, but It's pretty good for the most part.